pyproject.toml
[project]
name = "mathshooter"
version = "1.0.0"
description = "A math scrolling arcade shooter with modern Python patterns."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pygame-ce>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "pytest-mock>=3.10.0",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP"]
ignore = []

[tool.ruff.lint.isort]
known-first-party = ["assets", "config", "entities", "game", "interfaces", "operations", "ui"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
mypy_path = "src"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pygame.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
